Preserve route contracts while reducing roadmap weight
App stays the composition root while route ownership, layout shell, and deferred shell widgets move into local app modules. API route registration now uses a sequential default-router mount helper without changing public URLs, while special-case mounts remain explicit for contract visibility. Deal-pipeline validation and cursor helpers moved behind service-owned modules with the existing route schema export preserved.

Constraint: Public route URLs, provider order, v2 bypass behavior, route aliases, and dealPipelineValidationSchemas had to remain stable.

Rejected: Collapse route mounts into one table | source-pinned special mounts and telemetry boundaries need to stay explicit.

Rejected: Continue into lp-api or allocations extraction | requested slice 21 only starts with deal-pipeline and keeps scope bounded.

Confidence: high

Scope-risk: moderate

Directive: Do not start route-layer Pino migration or deeper DB-heavy route extraction without rerunning the targeted route-contract gates.

Tested: App route-governance 4 files / 55 tests; route-mounting 7 files / 42 tests; deal-pipeline contract 13 tests; taxonomy 3 tests; deal-pipeline API integration 3 active tests / 13 pre-existing skips; npm run check; npm run build:prod; npm run test:unit; npm run lint; npm run docs:routing:check; git diff --check

Not-tested: Full deal-pipeline API quarantine exits; existing Vite large chunk warning remains.
